Initial Version

+++ prog_odroid/project_detector/class_def.h
@@ -25,16 +25,16 @@ class Stream_in
 {
 	public:
 		uint16_t t_poll;
-		uint8_t max_num_buffer;
+		uint16_t max_num_buffer;
 		uint16_t width, height;
 		uint16_t offset_x, offset_y;
-		uint8_t binning;
+		uint16_t binning;
 		uint16_t expo_auto_max;
 		uint16_t grab_time_out;
 		float fov[2];
-		uint8_t max_no_detect;
-		uint8_t size_blur;
-		uint8_t bin_square;
+		uint16_t max_no_detect;
+		uint16_t size_blur;
+		uint16_t bin_square;
 		float k_square;
 		uint16_t area_square;
 		float cos_square;
@@ -52,8 +52,8 @@ class Stream_out
 	public:
 		float sub_angles[4];
 		uint64_t time_frame;
-		uint8_t no_detect;
-		uint8_t init_detect;
+		uint16_t no_detect;
+		uint16_t init_detect;
 
 		// Store stream_out data to buffer
 		void to_buffer(uint8_t * write_buffer) const;
